// frontend/lib/api/model/automationWithTriggersSchema.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { ConditionType } from './conditionType';
import type { ActionWithTriggersSchema } from './actionWithTriggersSchema';

export interface AutomationWithTriggersSchema {
  /**
   * @minimum -2147483648
   * @maximum 2147483647
   */
  id: number;
  /** @maxLength 255 */
  name: string;
  /**
   * @minimum -2147483648
   * @maximum 2147483647
   */
  on_value: number;
  /**
   * LT: lt<br/>GT: gt<br/>EQ: eq
   * @maxLength 2
   */
  condition: ConditionType;
  actions: ActionWithTriggersSchema[];
  /**
   * @minimum -2147483648
   * @maximum 2147483647
   */
  sensor_id: number;
  /**
   * @minimum -2147483648
   * @maximum 2147483647
   */
  user_id: number;
}